feat(security): expose certs without reqwest
