+++ security-utils/src/lib.rs
@@ -1,4 +1,6 @@
 #![forbid(unsafe_code)]
 
+pub mod certs;
+
 #[cfg(feature = "reqwest")]
 pub mod reqwest;
